React rewrite: moved tracks components and containers to own directories.
