+++ app/assets/javascripts/reactApp/components/tracks/List.js.jsx
@@ -1,12 +1,12 @@
 import React, { PropTypes, Component } from 'react';
-import SearchSchema   from '../schemas/search.js.es6';
-import TrackSchema    from '../schemas/track.js.es6';
-import TrackContainer from '../containers/track.js.jsx';
-import Spinner        from './spinner.js.jsx';
+import TracksListItemContainer         from '../../containers/tracks/ListItem.js.jsx';
+import TracksActions                   from '../../actions/tracks.js.es6';
+import SearchSchema                    from '../../schemas/search.js.es6';
+import TrackSchema                     from '../../schemas/track.js.es6';
+import Spinner                         from './../spinner.js.jsx';
 
-import TracksActions from '../actions/tracks.js.es6';
 
-class TrackList extends Component {
+class TracksList extends Component {
   componentDidMount() {
     if (!this.props.tracks.length && !this.props.isFetching) {
       this.props.dispatch(TracksActions.fetch(this.props.search));
@@ -42,7 +42,7 @@ class TrackList extends Component {
                 </div>
                 <div className="panel-body" data-hook="track-list">
                   {this.props.tracks.map(track =>
-                    <TrackContainer key={track.id} {...track} />
+                    <TracksListItemContainer key={track.id} {...track} />
                   )}
                 </div>
               </div>
@@ -54,11 +54,11 @@ class TrackList extends Component {
   }
 }
 
-TrackList.propTypes = {
+TracksList.propTypes = {
   tracks:     PropTypes.arrayOf(TrackSchema.PropTypes).isRequired,
   search:     SearchSchema.PropTypes.isRequired,
   dispatch:   PropTypes.func.isRequired,
   isFetching: PropTypes.bool.isRequired
 };
 
-export default TrackList;
+export default TracksList;
